Move RefOps trait to secret sharing module
It is too specific to be in the finite fields, basically nothing except `Linear` can conform to it
