+++ src/secret_sharing/scheme.rs
@@ -24,5 +24,34 @@ pub trait Linear<V: SharedValue>:
 {
 }
 
+/// The trait for arithmetic operations on references to a secret share, taking the
+/// second operand either by value or by reference. Secret sharings can be added/subtracted and
+/// multiplied by the shared values locally.
+///
+/// The need for this trait is dictated by [`rust-issue`] that causes us not being able to constrain
+/// references to `Self`. Once this issue is fixed, we can simply get rid of it.
+///
+/// This is automatically implemented for types which implement the operators. The idea is borrowed
+/// from [`RefNum`] trait, but I couldn't really make it work with HRTB and secret shares. Primitive
+/// types worked just fine though, so it is possible that it is another compiler issue.
+///
+/// [`RefNum`]: https://docs.rs/num/0.4.1/num/traits/trait.RefNum.html
+/// [`rust-issue`]: https://github.com/rust-lang/rust/issues/20671
+pub trait RefOps<'a, Base: 'a, R: 'a>:
+LocalArithmeticOps<Base, Base>
++ LocalArithmeticOps<&'a Base, Base>
++ Mul<R, Output = Base>
++ Mul<&'a R, Output = Base>
+{
+}
+impl<'a, T, Base: 'a, R: 'a> RefOps<'a, Base, R> for T where
+    T: LocalArithmeticOps<Base, Base>
+    + LocalArithmeticOps<&'a Base, Base>
+    + 'a
+    + Mul<R, Output = Base>
+    + Mul<&'a R, Output = Base>
+{
+}
+
 /// Secret share of a secret in bits. It has additive and multiplicative properties.
 pub trait Bitwise<V: GaloisField>: SecretSharing<V> + Linear<V> {}
